Spider Enemies now attack towers within range
- [Modified] EnemySpider script to attack and destroy towers within range.
- [Remaining] spiders can't continue walking on the path after coming back to 'Moving' state

+++ Assets/_TowerDefense3D/Scripts/World/Items/CannonAmmo.cs
@@ -25,7 +25,7 @@ namespace TowerDefense3D
             isSpawned = true;
         }
 
-        public override void DealDamage(IDamageDealer damageDealer, IDamageable defender, float damage, Vector3 hitPoint)
+        public override void DealDamage(IDamageDealer damageDealer, IDamageable defender, int damage, Vector3 hitPoint)
         {
             base.DealDamage(damageDealer, defender, damage, hitPoint);
         }
